requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-CORS==4.0.0
Pillow==10.0.0
